Not final

+++ lib/features/secretary_features/course/data/models/section_progress_model.dart
@@ -1,6 +1,6 @@
 class SectionProgressModel {
   final String message;
-  final int progressPercentage;
+  final double progressPercentage;
 
   SectionProgressModel({
     required this.message,
